requirements.txt
APScheduler==3.9.1
beautifulsoup4==4.11.2
colorama==0.4.4
img2pdf==0.4.4
Jinja2==3.1.2
opencv_python==4.6.0.66
pdfkit==1.0.0
pika==1.3.1
Pillow==9.4.0
PyAutoGUI==0.9.53
pycryptodomex==3.17
pydub==0.25.1
pymongo==4.2.0
PyMySQL==1.0.2
pyshadow==0.0.4
python-decouple==3.8
requests==2.25.1
selenium==4.8.2
SpeechRecognition==3.8.1
TwoCaptcha==0.0.1
undetected_chromedriver==3.4.6
weasyprint==58.0
